vue presque fini

+++ src/java/fr/tanchou/pvz/abstractEnity/abstractPlant/PassivePlant.java
@@ -1,7 +1,7 @@
 package fr.tanchou.pvz.abstractEnity.abstractPlant;
 
 public abstract class PassivePlant extends Plant {
-    public PassivePlant(int healthPoint, int colliderRadius, double x, int y,String name, int cost) {
+    public PassivePlant(int healthPoint, double colliderRadius, double x, int y,String name, int cost) {
         super(healthPoint,colliderRadius,x,y, name, cost);
     }
